#cele-27 Separate viewer settings drawer
